Revert "Arguments (#19)"
This reverts commit 6825cb716997d89d8c281c2cd9cd72b2b38b8b50.
