+++ build.gradle.kts
@@ -50,7 +50,6 @@ dependencies {
 
   // Use the Kotlin JUnit integration.
   testImplementation("org.jetbrains.kotlin:kotlin-test-junit")
-    implementation(kotlin("script-runtime"))
 }
 
 tasks.generateGrammarSource {
@@ -113,9 +112,9 @@ tasks.withType<GenerateModuleMetadata> {
   enabled = false
 }
 
-/*signing {
+signing {
   val signingKey: String? by project
   val signingPassword: String? by project
   useInMemoryPgpKeys(signingKey, signingPassword)
   sign(publishing.publications["maven"])
-}*/
+}
