+++ gradle.properties
@@ -1,1 +1,1 @@
-version=0.3-SNAPSHOT
+version=0.2
