+++ src/main/kotlin/io/github/oxisto/reticulated/PythonParser.kt
@@ -14,7 +14,6 @@ import org.antlr.v4.runtime.CommonTokenStream
 class PythonParser {
 
   fun parse(path: String): ParserResult {
-
     val inputStream = CharStreams.fromFileName(path)
     val lexer = Python3Lexer(inputStream)
     val tokenStream = CommonTokenStream(lexer)
@@ -30,3 +29,4 @@ class PythonParser {
     return ParserResult(fileInput, global)
   }
 }
+
