+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/Argument.kt
@@ -1,0 +1,11 @@
+package io.github.oxisto.reticulated.ast.expression
+
+import io.github.oxisto.reticulated.ast.Node
+
+class Argument(val expression: Expression) : Node() {
+
+  init {
+    expression.parent = this
+  }
+
+}
