+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/ArgumentList.kt
@@ -1,0 +1,11 @@
+package io.github.oxisto.reticulated.ast.expression
+
+import io.github.oxisto.reticulated.ast.Container
+import io.github.oxisto.reticulated.ast.Node
+import io.github.oxisto.reticulated.ast.Scope
+
+class ArgumentList(val arguments: List<Argument> = ArrayList()) : Node(), Container<Argument> {
+
+  override val children get() = this.arguments
+
+}
