+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/ArgumentListVisitor.kt
@@ -1,4 +1,4 @@
-package io.github.oxisto.reticulated.ast.expression.argument
+package io.github.oxisto.reticulated.ast.expression
 
 import io.github.oxisto.reticulated.ast.EmptyContextException
 import io.github.oxisto.reticulated.ast.Scope
@@ -7,29 +7,16 @@ import io.github.oxisto.reticulated.grammar.Python3Parser
 
 class ArgumentListVisitor(val scope: Scope) : Python3BaseVisitor<ArgumentList>() {
 
-  /**
-   * It is the trailer form the call in the form of: "(" [ argument_list [","] | comprehension ] ")"
-   * [see: {@linktourl https://docs.python.org/3/reference/expressions.html#calls }]
-   */
   override fun visitTrailer(ctx: Python3Parser.TrailerContext?): ArgumentList {
     if (ctx == null) {
       throw EmptyContextException()
     }
 
     return if (ctx.childCount == 3) {
+      // TODO: Handle comprehension
 
       // second child should be the argument list
-      val trailer = ctx.getChild(1)
-      if ( trailer.childCount == 1 ){
-
-        val argumentContext = trailer.getChild(0)
-        if ( argumentContext.childCount == 2 ) {
-          // The trailer is a comprehension
-          argumentContext.accept(ComprehensionVisitor(this.scope))
-        }
-      }
-      // The trailer is a argument_list
-      trailer.accept(this)
+      return ctx.getChild(1).accept(this)
     } else {
       ArgumentList()
     }
@@ -50,5 +37,4 @@ class ArgumentListVisitor(val scope: Scope) : Python3BaseVisitor<ArgumentList>()
 
     return ArgumentList(arguments)
   }
-
 }
