+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/Call.kt
@@ -1,7 +1,5 @@
 package io.github.oxisto.reticulated.ast.expression
 
-import io.github.oxisto.reticulated.ast.expression.argument.ArgumentList
-
 class Call(val primary: Primary, val argumentList: ArgumentList = ArgumentList()) : Primary() {
 
   init {
