+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/IdentifierVisitor.kt
@@ -7,18 +7,13 @@ import org.antlr.v4.runtime.tree.TerminalNode
 
 class IdentifierVisitor(val scope: Scope) : Python3BaseVisitor<Identifier>() {
 
-  /**
-   * see visitTerminal in the ExpressionVisitor
-   */
   override fun visitTerminal(node: TerminalNode?): Identifier {
     if (node == null) {
-      throw EmptyContextException()
+      throw EmptyContextException();
     }
 
     // TODO: literals and stuff
-    // TODO: functions without "primary." identifier are not included
-
-    // return super.visitTerm(ctx)
+    //return super.visitTerm(ctx)
 
     val id = Identifier(node.text)
