--- src/test/resources/comprehension_argument.py
@@ -1,5 +1,0 @@
-#/usr/bin/env python3
-
-def fun(a):
-    print(a)
-fun(x for x in range(10))
