+++ src/test/resources/import.py
@@ -1,5 +1,4 @@
 #/usr/bin/env python3
-
 import os
 
 print(os.name)
