--- src/test/resources/list_comprehension.py
@@ -1,3 +1,0 @@
-#/usr/bin/env python3
-
-nums = [x for x in range(10)]
