+++ src/test/resources/simple_stmt.py
@@ -1,12 +1,4 @@
 #!/usr/bin/env python3
 
 i = 4
-
-# x = y = 42    # has 5 children#
-# x = y = z = 42    # has 7 children
-# x = lambda y=72, z=2: y**7    # has 3 children
-# x, y, z = 1, 2, 3   # has 3 children
-# w = x if x < y else y if y < z else z   # has 3 children
-# print(w)  # has 1 children
-
 print(i)
